refactor(tests): address review feedback on virt fixtures
- Move rhel_vm_with_instance_type_and_preference and
  rhel_vm_with_cluster_instance_type_and_preference from instance_types.py
  to virtual_machines.py (all VM creation fixtures in one place)
- Extract ping_process_in_rhel_os into new vm_actions.py
- Move skip_if_no_cpumanager_workers and hugepages_gib_values back to
  tests/conftest.py
